requirements.txt
Flask
python-pptx
Pillow
gunicorn
